+++ RWAPIMicroservicePython/__init__.py
@@ -2,3 +2,4 @@ from RWAPIMicroservicePython.errors import NotFound, ApiKeyError, ValidationErro
 from RWAPIMicroservicePython.main import register, request_to_microservice
 from RWAPIMicroservicePython.cloudwatch import CloudWatchService
 
+from RWAPIMicroservicePython.test_utils import mock_request_validation