+++ RWAPIMicroservicePython/tests/__init__.py
@@ -1,2 +1,2 @@
-from RWAPIMicroservicePython.tests.constants import USER, APPLICATION
-from RWAPIMicroservicePython.tests.mocks import mock_request_validation, mock_request_validation_invalid_token
+from RWAPIMicroservicePython.test_utils import USER, APPLICATION, MICROSERVICE
+from RWAPIMicroservicePython.test_utils import mock_request_validation, mock_request_validation_invalid_token