+++ RWAPIMicroservicePython/tests/inject_logged_user_with_api_key_tests.py
@@ -11,7 +11,7 @@ from moto import mock_logs
 @requests_mock.mock(kw='mocker')
 @mock_logs()
 def test_inject_logged_user(mocker):
-    get_user_data_calls = mock_request_validation(mocker, application=None)
+    get_user_data_calls = mock_request_validation(mocker)
 
     test_endpoints = Blueprint('rw_api', __name__)
 
@@ -33,28 +33,28 @@ def test_inject_logged_user(mocker):
         app=app,
         gateway_url='https://gateway-url.com',
         token='microserviceToken',
-        require_api_key=False,
+        require_api_key=True,
         aws_region='us-east-1',
         aws_cloud_watch_log_stream_name='rw-api-microservice-python'
     )
 
-    response = app.test_client().get('/test', headers={'Authorization': 'Bearer abcd'})
+    response = app.test_client().get('/test', headers={'Authorization': 'Bearer abcd', 'x-api-key': 'api-key-test'})
     assert response.status_code == 200
     assert response.data == b'ok'
 
-    response = app.test_client().put('/test', headers={'Authorization': 'Bearer abcd'})
+    response = app.test_client().put('/test', headers={'Authorization': 'Bearer abcd', 'x-api-key': 'api-key-test'})
     assert response.status_code == 200
     assert response.data == b'ok'
 
-    response = app.test_client().post('/test', headers={'Authorization': 'Bearer abcd'})
+    response = app.test_client().post('/test', headers={'Authorization': 'Bearer abcd', 'x-api-key': 'api-key-test'})
     assert response.status_code == 200
     assert response.data == b'ok'
 
-    response = app.test_client().patch('/test', headers={'Authorization': 'Bearer abcd'})
+    response = app.test_client().patch('/test', headers={'Authorization': 'Bearer abcd', 'x-api-key': 'api-key-test'})
     assert response.status_code == 200
     assert response.data == b'ok'
 
-    response = app.test_client().delete('/test', headers={'Authorization': 'Bearer abcd'})
+    response = app.test_client().delete('/test', headers={'Authorization': 'Bearer abcd', 'x-api-key': 'api-key-test'})
     assert response.status_code == 200
     assert response.data == b'ok'
 
@@ -68,15 +68,15 @@ def test_inject_logged_user(mocker):
     )['events']
     assert len(log_lines) == 5
     assert log_lines[0][
-               'message'] == '{"request": {"method": "GET", "path": "/test", "query": {}}, "loggedUser": {"id": "1a10d7c6e0a37126611fd7a5", "name": "test user", "role": "USER", "provider": "local"}, "requestApplication": {"id": "anonymous", "name": "anonymous", "organization": null, "user": null, "apiKeyValue": null}}'
+               'message'] == '{"request": {"method": "GET", "path": "/test", "query": {}}, "loggedUser": {"id": "1a10d7c6e0a37126611fd7a5", "name": "test user", "role": "USER", "provider": "local"}, "requestApplication": {"id": "649c4b204967792f3a4e52c9", "name": "grouchy-armpit", "organization": null, "user": null, "apiKeyValue": "api-key-test"}}'
     assert log_lines[1][
-               'message'] == '{"request": {"method": "PUT", "path": "/test", "query": {}}, "loggedUser": {"id": "1a10d7c6e0a37126611fd7a5", "name": "test user", "role": "USER", "provider": "local"}, "requestApplication": {"id": "anonymous", "name": "anonymous", "organization": null, "user": null, "apiKeyValue": null}}'
+               'message'] == '{"request": {"method": "PUT", "path": "/test", "query": {}}, "loggedUser": {"id": "1a10d7c6e0a37126611fd7a5", "name": "test user", "role": "USER", "provider": "local"}, "requestApplication": {"id": "649c4b204967792f3a4e52c9", "name": "grouchy-armpit", "organization": null, "user": null, "apiKeyValue": "api-key-test"}}'
     assert log_lines[2][
-               'message'] == '{"request": {"method": "POST", "path": "/test", "query": {}}, "loggedUser": {"id": "1a10d7c6e0a37126611fd7a5", "name": "test user", "role": "USER", "provider": "local"}, "requestApplication": {"id": "anonymous", "name": "anonymous", "organization": null, "user": null, "apiKeyValue": null}}'
+               'message'] == '{"request": {"method": "POST", "path": "/test", "query": {}}, "loggedUser": {"id": "1a10d7c6e0a37126611fd7a5", "name": "test user", "role": "USER", "provider": "local"}, "requestApplication": {"id": "649c4b204967792f3a4e52c9", "name": "grouchy-armpit", "organization": null, "user": null, "apiKeyValue": "api-key-test"}}'
     assert log_lines[3][
-               'message'] == '{"request": {"method": "PATCH", "path": "/test", "query": {}}, "loggedUser": {"id": "1a10d7c6e0a37126611fd7a5", "name": "test user", "role": "USER", "provider": "local"}, "requestApplication": {"id": "anonymous", "name": "anonymous", "organization": null, "user": null, "apiKeyValue": null}}'
+               'message'] == '{"request": {"method": "PATCH", "path": "/test", "query": {}}, "loggedUser": {"id": "1a10d7c6e0a37126611fd7a5", "name": "test user", "role": "USER", "provider": "local"}, "requestApplication": {"id": "649c4b204967792f3a4e52c9", "name": "grouchy-armpit", "organization": null, "user": null, "apiKeyValue": "api-key-test"}}'
     assert log_lines[4][
-               'message'] == '{"request": {"method": "DELETE", "path": "/test", "query": {}}, "loggedUser": {"id": "1a10d7c6e0a37126611fd7a5", "name": "test user", "role": "USER", "provider": "local"}, "requestApplication": {"id": "anonymous", "name": "anonymous", "organization": null, "user": null, "apiKeyValue": null}}'
+               'message'] == '{"request": {"method": "DELETE", "path": "/test", "query": {}}, "loggedUser": {"id": "1a10d7c6e0a37126611fd7a5", "name": "test user", "role": "USER", "provider": "local"}, "requestApplication": {"id": "649c4b204967792f3a4e52c9", "name": "grouchy-armpit", "organization": null, "user": null, "apiKeyValue": "api-key-test"}}'
 
 
 @requests_mock.mock(kw='mocker')
@@ -100,7 +100,7 @@ def test_inject_microservice_user_no_api_key(mocker):
         app=app,
         gateway_url='https://gateway-url.com',
         token='microserviceToken',
-        require_api_key=False,
+        require_api_key=True,
         aws_region='us-east-1',
         aws_cloud_watch_log_stream_name='rw-api-microservice-python'
     )
@@ -148,8 +148,63 @@ def test_inject_microservice_user_no_api_key(mocker):
 
 @requests_mock.mock(kw='mocker')
 @mock_logs()
+def test_inject_logged_user_no_api_key(mocker):
+    get_user_data_calls = mock_request_validation(mocker, application=None)
+
+    test_endpoints = Blueprint('rw_api', __name__)
+
+    @test_endpoints.route('/test', methods=['GET', 'PUT', 'POST', 'PATCH', 'DELETE'])
+    def test_route():
+        return 'ok', 200
+
+    app = Flask(__name__)
+
+    app.register_blueprint(test_endpoints)
+
+    RWAPIMicroservicePython.register(
+        app=app,
+        gateway_url='https://gateway-url.com',
+        token='microserviceToken',
+        require_api_key=True,
+        aws_region='us-east-1',
+        aws_cloud_watch_log_stream_name='rw-api-microservice-python'
+    )
+
+    response = app.test_client().get('/test', headers={'Authorization': 'Bearer abcd'})
+    assert response.status_code == 403
+    assert response.data == b'{"errors":{"code":403,"message":"Required API key not found"}}\n'
+
+    response = app.test_client().put('/test', headers={'Authorization': 'Bearer abcd'})
+    assert response.status_code == 403
+    assert response.data == b'{"errors":{"code":403,"message":"Required API key not found"}}\n'
+
+    response = app.test_client().post('/test', headers={'Authorization': 'Bearer abcd'})
+    assert response.status_code == 403
+    assert response.data == b'{"errors":{"code":403,"message":"Required API key not found"}}\n'
+
+    response = app.test_client().patch('/test', headers={'Authorization': 'Bearer abcd'})
+    assert response.status_code == 403
+    assert response.data == b'{"errors":{"code":403,"message":"Required API key not found"}}\n'
+
+    response = app.test_client().delete('/test', headers={'Authorization': 'Bearer abcd'})
+    assert response.status_code == 403
+    assert response.data == b'{"errors":{"code":403,"message":"Required API key not found"}}\n'
+
+    assert get_user_data_calls.called
+    assert get_user_data_calls.call_count == 5
+
+    aws_mock = boto3.client('logs', region_name='us-east-1')
+    log_lines = aws_mock.get_log_events(
+        logGroupName="api-keys-usage",
+        logStreamName="rw-api-microservice-python"
+    )['events']
+    assert len(log_lines) == 0
+
+
+@requests_mock.mock(kw='mocker')
+@mock_logs()
 def test_inject_logged_user_when_no_authorization_header_is_present(mocker):
-    get_user_data_calls = mock_request_validation(mocker, user=None, application=None)
+    get_user_data_calls = mock_request_validation(mocker, user=None)
 
     test_endpoints = Blueprint('rw_api', __name__)
 
@@ -167,7 +222,7 @@ def test_inject_logged_user_when_no_authorization_header_is_present(mocker):
         app=app,
         gateway_url='https://gateway-url.com',
         token='microserviceToken',
-        require_api_key=False,
+        require_api_key=True,
         aws_region='us-east-1',
         aws_cloud_watch_log_stream_name='rw-api-microservice-python'
     )
@@ -202,15 +257,15 @@ def test_inject_logged_user_when_no_authorization_header_is_present(mocker):
     )['events']
     assert len(log_lines) == 5
     assert log_lines[0][
-               'message'] == '{"request": {"method": "GET", "path": "/test", "query": {}}, "loggedUser": {"id": "anonymous", "name": "anonymous", "role": "anonymous", "provider": "anonymous"}, "requestApplication": {"id": "anonymous", "name": "anonymous", "organization": null, "user": null, "apiKeyValue": null}}'
+               'message'] == '{"request": {"method": "GET", "path": "/test", "query": {}}, "loggedUser": {"id": "anonymous", "name": "anonymous", "role": "anonymous", "provider": "anonymous"}, "requestApplication": {"id": "649c4b204967792f3a4e52c9", "name": "grouchy-armpit", "organization": null, "user": null, "apiKeyValue": "api-key-test"}}'
     assert log_lines[1][
-               'message'] == '{"request": {"method": "PUT", "path": "/test", "query": {}}, "loggedUser": {"id": "anonymous", "name": "anonymous", "role": "anonymous", "provider": "anonymous"}, "requestApplication": {"id": "anonymous", "name": "anonymous", "organization": null, "user": null, "apiKeyValue": null}}'
+               'message'] == '{"request": {"method": "PUT", "path": "/test", "query": {}}, "loggedUser": {"id": "anonymous", "name": "anonymous", "role": "anonymous", "provider": "anonymous"}, "requestApplication": {"id": "649c4b204967792f3a4e52c9", "name": "grouchy-armpit", "organization": null, "user": null, "apiKeyValue": "api-key-test"}}'
     assert log_lines[2][
-               'message'] == '{"request": {"method": "POST", "path": "/test", "query": {}}, "loggedUser": {"id": "anonymous", "name": "anonymous", "role": "anonymous", "provider": "anonymous"}, "requestApplication": {"id": "anonymous", "name": "anonymous", "organization": null, "user": null, "apiKeyValue": null}}'
+               'message'] == '{"request": {"method": "POST", "path": "/test", "query": {}}, "loggedUser": {"id": "anonymous", "name": "anonymous", "role": "anonymous", "provider": "anonymous"}, "requestApplication": {"id": "649c4b204967792f3a4e52c9", "name": "grouchy-armpit", "organization": null, "user": null, "apiKeyValue": "api-key-test"}}'
     assert log_lines[3][
-               'message'] == '{"request": {"method": "PATCH", "path": "/test", "query": {}}, "loggedUser": {"id": "anonymous", "name": "anonymous", "role": "anonymous", "provider": "anonymous"}, "requestApplication": {"id": "anonymous", "name": "anonymous", "organization": null, "user": null, "apiKeyValue": null}}'
+               'message'] == '{"request": {"method": "PATCH", "path": "/test", "query": {}}, "loggedUser": {"id": "anonymous", "name": "anonymous", "role": "anonymous", "provider": "anonymous"}, "requestApplication": {"id": "649c4b204967792f3a4e52c9", "name": "grouchy-armpit", "organization": null, "user": null, "apiKeyValue": "api-key-test"}}'
     assert log_lines[4][
-               'message'] == '{"request": {"method": "DELETE", "path": "/test", "query": {}}, "loggedUser": {"id": "anonymous", "name": "anonymous", "role": "anonymous", "provider": "anonymous"}, "requestApplication": {"id": "anonymous", "name": "anonymous", "organization": null, "user": null, "apiKeyValue": null}}'
+               'message'] == '{"request": {"method": "DELETE", "path": "/test", "query": {}}, "loggedUser": {"id": "anonymous", "name": "anonymous", "role": "anonymous", "provider": "anonymous"}, "requestApplication": {"id": "649c4b204967792f3a4e52c9", "name": "grouchy-armpit", "organization": null, "user": null, "apiKeyValue": "api-key-test"}}'
 
 
 @requests_mock.mock(kw='mocker')
@@ -232,32 +287,32 @@ def test_inject_logged_user_when_token_is_invalid(mocker):
         app=app,
         gateway_url='https://gateway-url.com',
         token='microserviceToken',
-        require_api_key=False,
+        require_api_key=True,
         aws_region='us-east-1',
         aws_cloud_watch_log_stream_name='rw-api-microservice-python'
     )
 
-    response = app.test_client().get('/test', headers={'Authorization': 'Bearer abcd'})
+    response = app.test_client().get('/test', headers={'Authorization': 'Bearer abcd', 'x-api-key': 'api-key-test'})
     assert response.status_code == 401
     assert response.json == json.loads(
         b'{"errors": [{"status": 401, "detail": "Your token is outdated. Please use /auth/login to login and /auth/generate-token to generate a new token."}]}')
 
-    response = app.test_client().put('/test', headers={'Authorization': 'Bearer abcd'})
+    response = app.test_client().put('/test', headers={'Authorization': 'Bearer abcd', 'x-api-key': 'api-key-test'})
     assert response.status_code == 401
     assert response.json == json.loads(
         b'{"errors": [{"status": 401, "detail": "Your token is outdated. Please use /auth/login to login and /auth/generate-token to generate a new token."}]}')
 
-    response = app.test_client().post('/test', headers={'Authorization': 'Bearer abcd'})
+    response = app.test_client().post('/test', headers={'Authorization': 'Bearer abcd', 'x-api-key': 'api-key-test'})
     assert response.status_code == 401
     assert response.json == json.loads(
         b'{"errors": [{"status": 401, "detail": "Your token is outdated. Please use /auth/login to login and /auth/generate-token to generate a new token."}]}')
 
-    response = app.test_client().patch('/test', headers={'Authorization': 'Bearer abcd'})
+    response = app.test_client().patch('/test', headers={'Authorization': 'Bearer abcd', 'x-api-key': 'api-key-test'})
     assert response.status_code == 401
     assert response.json == json.loads(
         b'{"errors": [{"status": 401, "detail": "Your token is outdated. Please use /auth/login to login and /auth/generate-token to generate a new token."}]}')
 
-    response = app.test_client().delete('/test', headers={'Authorization': 'Bearer abcd'})
+    response = app.test_client().delete('/test', headers={'Authorization': 'Bearer abcd', 'x-api-key': 'api-key-test'})
     assert response.status_code == 401
     assert response.json == json.loads(
         b'{"errors": [{"status": 401, "detail": "Your token is outdated. Please use /auth/login to login and /auth/generate-token to generate a new token."}]}')