+++ RWAPIMicroservicePython/tests/request_to_microservice_tests.py
@@ -19,13 +19,13 @@ def request_has_body(expected_body):
 
 @requests_mock.mock(kw='mocker')
 def test_get_request_to_microservice_happy_case(mocker):
-    get_calls = mocker.get('http://gateway-url.com/v1/microservice/endpoint',
+    get_calls = mocker.get('/v1/microservice/endpoint',
                            request_headers={'x-api-key': 'api-key-test'},
                            additional_matcher=request_has_no_body,
                            json={})
 
     RWAPIMicroservicePython.MICROSERVICE_TOKEN = 'microserviceToken'
-    RWAPIMicroservicePython.GATEWAY_URL = 'http://gateway-url.com'
+    RWAPIMicroservicePython.GATEWAY_URL = 'https://gateway-url.com'
 
     response = RWAPIMicroservicePython.request_to_microservice(
         method='GET',
@@ -41,12 +41,12 @@ def test_get_request_to_microservice_happy_case(mocker):
 
 @requests_mock.mock(kw='mocker')
 def test_delete_request_to_microservice_happy_case(mocker):
-    get_calls = mocker.delete('http://gateway-url.com/v1/microservice/endpoint',
+    get_calls = mocker.delete('/v1/microservice/endpoint',
                               request_headers={'x-api-key': 'api-key-test'}, additional_matcher=request_has_no_body,
                               json={})
 
     RWAPIMicroservicePython.MICROSERVICE_TOKEN = 'microserviceToken'
-    RWAPIMicroservicePython.GATEWAY_URL = 'http://gateway-url.com'
+    RWAPIMicroservicePython.GATEWAY_URL = 'https://gateway-url.com'
 
     response = RWAPIMicroservicePython.request_to_microservice(
         method='DELETE',
@@ -62,12 +62,12 @@ def test_delete_request_to_microservice_happy_case(mocker):
 
 @requests_mock.mock(kw='mocker')
 def test_patch_request_to_microservice_happy_case(mocker):
-    get_calls = mocker.patch('http://gateway-url.com/v1/microservice/endpoint',
+    get_calls = mocker.patch('/v1/microservice/endpoint',
                              request_headers={'x-api-key': 'api-key-test'},
                              additional_matcher=request_has_body({'key': 'value'}), json={})
 
     RWAPIMicroservicePython.MICROSERVICE_TOKEN = 'microserviceToken'
-    RWAPIMicroservicePython.GATEWAY_URL = 'http://gateway-url.com'
+    RWAPIMicroservicePython.GATEWAY_URL = 'https://gateway-url.com'
 
     response = RWAPIMicroservicePython.request_to_microservice(
         method='PATCH',
@@ -84,12 +84,12 @@ def test_patch_request_to_microservice_happy_case(mocker):
 
 @requests_mock.mock(kw='mocker')
 def test_put_request_to_microservice_happy_case(mocker):
-    get_calls = mocker.put('http://gateway-url.com/v1/microservice/endpoint',
+    get_calls = mocker.put('/v1/microservice/endpoint',
                            request_headers={'x-api-key': 'api-key-test'},
                            additional_matcher=request_has_body({'key': 'value'}), json={})
 
     RWAPIMicroservicePython.MICROSERVICE_TOKEN = 'microserviceToken'
-    RWAPIMicroservicePython.GATEWAY_URL = 'http://gateway-url.com'
+    RWAPIMicroservicePython.GATEWAY_URL = 'https://gateway-url.com'
 
     response = RWAPIMicroservicePython.request_to_microservice(
         method='PUT',
@@ -106,12 +106,12 @@ def test_put_request_to_microservice_happy_case(mocker):
 
 @requests_mock.mock(kw='mocker')
 def test_post_request_to_microservice_happy_case(mocker):
-    get_calls = mocker.post('http://gateway-url.com/v1/microservice/endpoint',
+    get_calls = mocker.post('/v1/microservice/endpoint',
                             request_headers={'x-api-key': 'api-key-test'},
                             additional_matcher=request_has_body({'key': 'value'}), json={})
 
     RWAPIMicroservicePython.MICROSERVICE_TOKEN = 'microserviceToken'
-    RWAPIMicroservicePython.GATEWAY_URL = 'http://gateway-url.com'
+    RWAPIMicroservicePython.GATEWAY_URL = 'https://gateway-url.com'
 
     response = RWAPIMicroservicePython.request_to_microservice(
         method='POST',
@@ -128,11 +128,11 @@ def test_post_request_to_microservice_happy_case(mocker):
 
 @requests_mock.mock(kw='mocker')
 def test_request_to_microservice_no_api_version(mocker):
-    get_calls = mocker.get('http://gateway-url.com/microservice/endpoint',
+    get_calls = mocker.get('/microservice/endpoint',
                            request_headers={'x-api-key': 'api-key-test'}, json={})
 
     RWAPIMicroservicePython.MICROSERVICE_TOKEN = 'microserviceToken'
-    RWAPIMicroservicePython.GATEWAY_URL = 'http://gateway-url.com'
+    RWAPIMicroservicePython.GATEWAY_URL = 'https://gateway-url.com'
 
     response = RWAPIMicroservicePython.request_to_microservice(
         method='GET',
@@ -148,7 +148,7 @@ def test_request_to_microservice_no_api_version(mocker):
 
 @requests_mock.mock(kw='mocker')
 def test_request_to_microservice_ct_404(mocker):
-    get_calls = mocker.get('http://gateway-url.com/v1/microservice/endpoint',
+    get_calls = mocker.get('/v1/microservice/endpoint',
                            request_headers={'x-api-key': 'api-key-test'},
                            status_code=404,
                            json={
@@ -161,7 +161,7 @@ def test_request_to_microservice_ct_404(mocker):
                            })
 
     RWAPIMicroservicePython.MICROSERVICE_TOKEN = 'microserviceToken'
-    RWAPIMicroservicePython.GATEWAY_URL = 'http://gateway-url.com'
+    RWAPIMicroservicePython.GATEWAY_URL = 'https://gateway-url.com'
 
     response = RWAPIMicroservicePython.request_to_microservice(
         method='GET',
@@ -177,11 +177,11 @@ def test_request_to_microservice_ct_404(mocker):
 
 @requests_mock.mock(kw='mocker')
 def test_request_to_microservice_http_404(mocker):
-    get_calls = mocker.get('http://gateway-url.com/v1/microservice/endpoint',
+    get_calls = mocker.get('/v1/microservice/endpoint',
                            request_headers={'x-api-key': 'api-key-test'}, status_code=404)
 
     RWAPIMicroservicePython.MICROSERVICE_TOKEN = 'microserviceToken'
-    RWAPIMicroservicePython.GATEWAY_URL = 'http://gateway-url.com'
+    RWAPIMicroservicePython.GATEWAY_URL = 'https://gateway-url.com'
 
     with pytest.raises(NotFound) as e:
         assert RWAPIMicroservicePython.request_to_microservice(